Feature worker (#295)
* fixing powershell freezing

* added an interim preconfig script

* forgot to call the config function lol

* disabling firewall using set-netfirewallprofile

* installer changed to be consistent with new configureVM module

* PS prompt fix + helper functions

* PS prompt backticks

---------

+++ Modules/configureVM.psm1
@@ -1,4 +1,5 @@
-# COMMANDO-VM DEBLOATER
+# COMMANDO-VM CONFIG
+$global:helpersExecuted = $false
 
 function Commando-Remove-App {
 # Function for removing Apps
@@ -17,13 +18,13 @@ function Commando-Remove-App {
             $result = Remove-AppxPackage -Package $packageFullName -ErrorAction Stop
 
             if ($null -eq $result) {
-                Write-Output "[DEBLOAT] Installed $appName has been successfully removed."
+                Write-Output "[+] Installed $appName has been successfully removed."
             } else {
-                Write-Output "[DEBLOAT] Failed to remove installed app $appName."
+                Write-Output "[+] Failed to remove installed app $appName."
             }
         }
         else {
-            Write-Output "[DEBLOAT] Installed $appName not found on the system."
+            Write-Output "[+] Installed $appName not found on the system."
         }
         # Check if the app is provisioned
         $provisionedPackage = Get-AppxProvisionedPackage -Online | Where-Object { $_.DisplayName -eq $appName } -ErrorAction SilentlyContinue
@@ -31,12 +32,12 @@ function Commando-Remove-App {
             $result = Remove-AppxProvisionedPackage -PackageName $provisionedPackage.PackageName -Online  
 
             if ($result) {
-                Write-Output "[DEBLOAT] Provisioned $appName has been successfully removed."
+                Write-Output "[+] Provisioned $appName has been successfully removed."
             } else {
-                Write-Output "[DEBLOAT] Failed to remove porvisioned app $appName."
+                Write-Output "[+] Failed to remove porvisioned app $appName."
             }
         } else {
-            Write-Output "[DEBLOAT] Provisioned $appName not found on the system."
+            Write-Output "[+] Provisioned $appName not found on the system."
         }
     } 
     catch {
@@ -58,17 +59,18 @@ function Commando-Remove-Service {
 
         if ($service) {
             $service | Set-Service -StartupType Manual -ErrorAction Stop
-            Write-Output "[DEBLOAT] Service $serviceName has been disabled."
+            Write-Output "[+] Service $serviceName has been disabled."
         } else {
-            Write-Output "[DEBLOAT] Service $serviceName not found."
+            Write-Output "[+] Service $serviceName not found."
         }
     }
     catch {
-        Write-Error "[DEBLOAT] An error occurred while setting the service startup type. Error: $_"
+        Write-Error "An error occurred while setting the service startup type. Error: $_"
     }
 }
 
 function Commando-Delete-Task {
+# Function for disabling scheduled tasks
     param(
         [Parameter(Mandatory=$true)]
         [ValidateNotNullOrEmpty()]
@@ -82,10 +84,10 @@ function Commando-Delete-Task {
     try {
         $output = Disable-ScheduledTask -TaskName $value -ErrorAction SilentlyContinue
         if ($output){
-            Write-Output "[DEBLOAT] Scheduled task '$name' has been disabled."
+            Write-Output "[+] Scheduled task '$name' has been disabled."
         }
         else{
-            Write-Output "[DEBLOAT] Scheduled task '$name' not found."
+            Write-Output "[+] Scheduled task '$name' not found."
         }
     
     }
@@ -95,7 +97,7 @@ function Commando-Delete-Task {
 }
 
 function Commando-Remove-RegistryValue {
-    # Function for setting Registry items
+# Function for setting Registry items
     param(
         [Parameter(Mandatory=$true)]
         [ValidateNotNullOrEmpty()]
@@ -124,7 +126,8 @@ function Commando-Remove-RegistryValue {
             $validatedData = [int64]::Parse($data)
         }
         elseif ($type -eq "Binary") {
-            $validatedData = [byte[]]($data -split " " | % { [convert]::ToByte($_, 16) })
+            $validatedData = [byte[]]::new(($data -split '(.{2})' | Where-Object { $_ -match '..' } | ForEach-Object { [convert]::ToByte($_, 16) }))
+
         }
         else {
             $validatedData = $data
@@ -134,14 +137,14 @@ function Commando-Remove-RegistryValue {
         if (!(Test-Path -Path $path)) {
             # Create the registry key
             New-Item -Path $path -Force | Out-Null
-            Write-Output "Registry key created: $path"
+            Write-Output "`t[+] Registry key created: $path"
         }
         else {
-            Write-Output "Registry key already exists: $path"
+            Write-Output "`t[+] Registry key already exists: $path"
         }
 
         Set-ItemProperty -Path $path -Name $value -Value $validatedData -Type $type -Force | Out-Null
-        Write-Output "[DEBLOAT] $name has been successfully updated."
+        Write-Output "[+] $name has been successful"
     }
     catch {
         Write-Error "Failed to update the registry value. Error: $_"
@@ -168,17 +171,17 @@ function Commando-Remove-Path {
         if ($type -eq "file") {
             if (Test-Path -Path $path -PathType Leaf) {
                 Remove-Item -Path $path -Force
-                Write-Output "[DEBLOAT] $name has been successfully removed."
+                Write-Output "[+] $name has been successfully removed."
             } else {
-                Write-Output "[DEBLOAT] $path does not exist."
+                Write-Output "[+] $path does not exist."
             }
         }
         elseif ($type -eq "dir") {
             if (Test-Path -Path $path -PathType Container) {
                 Remove-Item -Path $path -Recurse -Force
-                Write-Output "[DEBLOAT] $name has been successfully removed."
+                Write-Output "[+] $name has been successfully removed."
             } else {
-                Write-Output "[DEBLOAT] $path does not exist."
+                Write-Output "[+] $path does not exist."
             }
         }
     }
@@ -201,30 +204,76 @@ function Commando-Remove-Custom{
     )
 
     try {
-        Write-Output "[DEBLOAT] Executing commands for '$name':"
+        Write-Output "[+] Executing commands for '$name':"
         foreach ($cmd in $cmds) {
-            Write-Output "`tExecuting command: $cmd"
+            Write-Output "`t[+] Executing command: $cmd"
             start-process powershell -ArgumentList "-WindowStyle","Hidden","-Command",$cmd -Wait
-            Write-Host "`tProcess completed. Moving to next."
+            Write-Host "`t[+] Process completed. Moving to next."
         }
-        Write-Output "[DEBLOAT] All commands for '$name' have been executed successfully."
+        Write-Output "[+] All commands for '$name' have been executed successfully."
     }
     catch {
         Write-Error "An error occurred while executing commands for '$name'. Error: $_"
     }
 }
 
+function Commando-Prompt {
+    $psprompt = @"
+        function prompt {
+            Write-Host ("COMMANDO " + `$(Get-Date)) -ForegroundColor Red
+            Write-Host ("PS " + `$(Get-Location) + " >") -NoNewLine -ForegroundColor White
+            return " "
+        }
+"@
 
-# DEBLOATER MAIN
-function Commando-Debloat {
+    # Ensure profile file exists and append new content to it, not overwriting old content
+    if (!(Test-Path $profile)) {
+        New-Item -ItemType File -Path $profile -Force | Out-Null
+    }
+    Add-Content -Path $profile -Value $psprompt
+
+    # Add timestamp to cmd prompt
+    Invoke-Expression ([System.Text.Encoding]::UTF8.GetString([System.Convert]::FromBase64String("Y21kIC9jICdzZXR4IFBST01QVCBDT01NQU5ETyRTJGQkcyR0JF8kcCQrJGcn"))) | Out-Null
+
+    Write-Host "[+] Timestamps added to cmd prompt and PowerShell" -ForegroundColor Green
+}
+
+function Commando-Logging {
+    if ($PSVersionTable -And $PSVersionTable.PSVersion.Major -ge 5) {
+        Write-Host "[+] Enabling PowerShell Script Block Logging" -ForegroundColor Green
+
+        $psLoggingPath = 'HKLM:\SOFTWARE\Wow6432Node\Policies\Microsoft\Windows\PowerShell'
+        if (-Not (Test-Path $psLoggingPath)) {
+            New-Item -Path $psLoggingPath -Force | Out-Null
+        }
+
+        $psLoggingPath = 'HKLM:\SOFTWARE\Wow6432Node\Policies\Microsoft\Windows\PowerShell\Transcription'
+        if (-Not (Test-Path $psLoggingPath)) {
+            New-Item -Path $psLoggingPath -Force | Out-Null
+        }
+        New-ItemProperty -Path $psLoggingPath -Name "EnableInvocationHeader" -Value 1 -PropertyType DWORD -Force | Out-Null
+        New-ItemProperty -Path $psLoggingPath -Name "EnableTranscripting" -Value 1 -PropertyType DWORD -Force | Out-Null
+        New-ItemProperty -Path $psLoggingPath -Name "OutputDirectory" -Value (Join-Path ${Env:UserProfile} "Desktop\PS_Transcripts") -PropertyType String -Force | Out-Null
+
+        $psLoggingPath = 'HKLM:\SOFTWARE\Wow6432Node\Policies\Microsoft\Windows\PowerShell\ScriptBlockLogging'
+        if (-Not (Test-Path $psLoggingPath)) {
+            New-Item -Path $psLoggingPath -Force | Out-Null
+        }
+        New-ItemProperty -Path $psLoggingPath -Name "EnableScriptBlockLogging" -Value 1 -PropertyType DWORD -Force | Out-Null
+        Write-Host "`t[i] PowerShell transcripts will be saved to the desktop." -ForegroundColor Green
+    }
+}
+
+# WORKER MAIN
+function Commando-Configure {
     param(
         [Parameter(Position = 0)]
-        [string]$debloatConfig = (Join-Path -Path $PSScriptRoot -ChildPath "debloatConfig.xml")
+        [string]$configFile = (Join-Path -Path $PSScriptRoot -ChildPath "debloatConfig.xml")
     )
 
     try {
         # Load and parse the XML config file
-        $config = [xml](Get-Content $debloatConfig)
+        $config = [xml](Get-Content $configFile)
 
         # Process the apps
         if ($config.config.apps.app) {
@@ -283,17 +332,26 @@ function Commando-Debloat {
             }
         }
 
-        # Unpinning all Start Tiles
-        Write-Output "[DEBLOAT] Unpinning all Start Menu Tiles."
-        $key = Get-ItemProperty -Path "HKCU:\Software\Microsoft\Windows\CurrentVersion\CloudStore\Store\Cache\DefaultAccount\*start.tilegrid`$windows.data.curatedtilecollection.tilecollection\Current"
-		$data = $key.Data[0..25] + ([byte[]](202,50,0,226,44,1,1,0,0))
-		Set-ItemProperty -Path $key.PSPath -Name "Data" -Type Binary -Value $data
-		Stop-Process -Name "ShellExperienceHost" -Force -ErrorAction SilentlyContinue
+        if (!$global:helpersExecuted) {
+            # TODO Needs to not run twice -- maybe global variable (?) but thats ugly
+            # Unpinning all Start Tiles
+            Write-Output "[+] Unpinning all Start Menu Tiles."
+            $key = Get-ItemProperty -Path "HKCU:\Software\Microsoft\Windows\CurrentVersion\CloudStore\Store\Cache\DefaultAccount\*start.tilegrid`$windows.data.curatedtilecollection.tilecollection\Current"
+            $data = $key.Data[0..25] + ([byte[]](202,50,0,226,44,1,1,0,0))
+            Set-ItemProperty -Path $key.PSPath -Name "Data" -Type Binary -Value $data
+            Stop-Process -Name "ShellExperienceHost" -Force -ErrorAction SilentlyContinue
+
+            Commando-Prompt
+            Commando-Logging
+
+            # Set the flag to indicate the operation has been executed
+            $global:helpersExecuted = $true
+        }
     }
     catch {
-        Write-Error "An error occurred while applying debloat. Error: $_"
+        Write-Error "An error occurred while applying config. Error: $_"
     }
 }
 
 # Export Function
-Export-ModuleMember -Function Commando-Debloat
+Export-ModuleMember -Function Commando-Configure
